feat: Add "+" button to FoodMenu component for each food item. Refactor ShowFoodMenu screen to render both FoodRibbon and FoodMenu.
