+++ src/components/FoodMenu.jsx
@@ -1,52 +1,53 @@
 import PropTypes from "prop-types";
 import { useSelector, useDispatch } from "react-redux";
 import { useNavigate, useParams } from "react-router-dom";
-// Importing the action creator for selecting a food item
 import { selectFood } from "../slices/foodDataApiSlice";
+import { addToCart } from "../slices/cartSlice"; // Import addToCart action
 
-
-const ShowFoodMenu = () => {
-  // Hooks for dispatching actions, navigating programmatically, and accessing URL parameters
+const FoodMenu = () => {
   const dispatch = useDispatch();
   const navigate = useNavigate();
-  const { category } = useParams(); // Extracting the 'category' parameter from the URL
+  const { category } = useParams();
 
-  // Fetching food data from the Redux store and filtering based on the current category
   const foodData = useSelector((state) => state.foodData.data);
-  const foods = foodData[category] || []; // Default to an empty array if the category is not found
+  const foods = foodData[category] || [];
 
-  // Handler function for clicking on a food item
   const handleFoodClick = (food) => {
-    dispatch(selectFood(food)); // Dispatching an action to select the clicked food item
-    navigate(`/food/${food.id}`); // Navigating to the food detail page using the food item's ID
+    dispatch(selectFood(food));
+    navigate(`/food/${food.id}`);
   };
 
-  // Rendering the food menu UI
+  const handleAddToCart = (food, e) => {
+    e.stopPropagation(); // Prevents triggering the handleFoodClick when adding to cart
+    dispatch(addToCart({ ...food, qty: 1 })); // Add food item to cart with initial quantity of 1
+  };
+
   return (
     <div className="cursor-pointer sm:p-2 sm:hover:shadow-slate-400 sm:shadow-md sm:border sm:border-slate-400 sm:m-1 transition-shadow duration-200 group">
       <h1>Food Menu</h1>
       {foods.map((food) => (
         <div key={food.id} onClick={() => handleFoodClick(food)}>
           <img
-            src={food.img} // Displaying the food item's image
+            src={food.img}
             width={342}
             height={542}
-            alt={food.name} // Providing an alt text for the image
+            alt={food.name}
             className="group-hover:opacity-80 transition-opacity duration-200"
-            style={{ maxWidth: "100%", height: "auto" }} // Ensuring the image is responsive
+            style={{ maxWidth: "100%", height: "auto" }}
           />
           <h4>{food.name}</h4>
           <p>{food.price}</p>
+          <button onClick={(e) => handleAddToCart(food, e)} className="add-to-cart-button">
+            +
+          </button>
         </div>
       ))}
     </div>
   );
 };
 
-// Defining PropTypes for the component to ensure proper usage
-ShowFoodMenu.propTypes = {
-  selectedCategory: PropTypes.string, // This prop seems to be unused and might be a leftover or mistake
+FoodMenu.propTypes = {
+  selectedCategory: PropTypes.string,
 };
 
-// Exporting the component for use in other parts of the application
-export default ShowFoodMenu;
+export default FoodMenu;
